[merge] Move options code to merge.options
